More code removals
This code will be restored in brewkit or other tea primitives
